+++ gameLoop.h
@@ -1,0 +1,9 @@
+#ifndef GAMELOOP_H
+#define GAMELOOP_H
+
+void aiTurn(int aiPlayer);
+void playerTurn(int currentPlayer);
+
+void announceWinner();
+
+#endif //GAMELOOP_H
