+++ menu.h
@@ -1,0 +1,10 @@
+#ifndef MENU_H
+#define MENU_H
+
+void startSinglePlayerGame();
+void startTwoPlayerGame();
+void startAIVsAITest();
+void showOptions();
+void showAbout();
+
+#endif //MENU_H
